# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RoadSoS Kids Emergency Response System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /incidents:
    get:
      operationId: listIncidents
      tags: [incidents]
      summary: List all incidents
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of incidents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Incident"
    post:
      operationId: createIncident
      tags: [incidents]
      summary: Create a new emergency incident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /incidents/{incidentId}:
    get:
      operationId: getIncident
      tags: [incidents]
      summary: Get incident by ID
      parameters:
        - name: incidentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Incident details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
    patch:
      operationId: updateIncidentStatus
      tags: [incidents]
      summary: Update incident status
      parameters:
        - name: incidentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentStatusUpdate"
      responses:
        "200":
          description: Updated incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /incidents/{incidentId}/vitals:
    get:
      operationId: getIncidentVitals
      tags: [vitals]
      summary: Get latest vitals for an incident
      parameters:
        - name: incidentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Vitals readings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VitalsReading"
    post:
      operationId: submitVitals
      tags: [vitals]
      summary: Submit new vitals reading from innerwear
      parameters:
        - name: incidentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VitalsInput"
      responses:
        "201":
          description: Vitals saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VitalsReading"
...
  /ambulances/nearest:
    get:
      operationId: getNearestAmbulance
      tags: [ambulances]
      summary: Find nearest available ambulance with ETA
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lon
          in: query
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Nearest ambulance with routing info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NearestAmbulanceResult"
        "404":
          description: No available ambulances
  /ambulances/{ambulanceId}/location:
    patch:
      operationId: updateAmbulanceLocation
      tags: [ambulances]
      summary: Update ambulance GPS location
      parameters:
        - name: ambulanceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ambulance"
...
  /innerwear/sensor-data:
    post:
      operationId: submitSensorData
      tags: [incidents]
      summary: Submit raw sensor data from innerwear device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SensorDataInput"
      responses:
        "200":
          description: Processed result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SensorDataResult"
...
components:
  schemas:
...
    Incident:
      type: object
      properties:
        id:
          type: string
        childId:
          type: string
        childName:
          type: string
        childAge:
          type: integer
        severity:
          type: string
        status:
          type: string
        incidentType:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        locationAddress:
          type: string
        accuracyMeters:
          type: number
        heartRate:
          type: ["integer", "null"]
        temperature:
          type: ["number", "null"]
        vitalsConfidence:
          type: ["number", "null"]
        heartRateTrend:
          type: ["string", "null"]
        parentName:
          type: string
        parentPhone:
          type: string
        parentStatus:
          type: string
        ambulanceId:
          type: ["string", "null"]
        ambulanceUnit:
          type: ["string", "null"]
        ambulanceEtaMinutes:
          type: ["integer", "null"]
        ambulanceDistanceKm:
          type: ["number", "null"]
        hospitalId:
          type: ["string", "null"]
        hospitalName:
          type: ["string", "null"]
        impactMagnitude:
          type: ["number", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        resolvedAt:
          type: ["string", "null"]
      required:
        - id
        - childId
        - childName
        - childAge
        - severity
        - status
        - incidentType
        - latitude
        - longitude
        - parentName
        - parentPhone
        - parentStatus
        - createdAt
        - updatedAt
    IncidentInput:
      type: object
      properties:
        childId:
          type: string
        severity:
          type: string
        incidentType:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        locationAddress:
          type: string
        impactMagnitude:
          type: number
        heartRate:
          type: integer
        temperature:
          type: number
      required: [childId, severity, incidentType, latitude, longitude]
    IncidentStatusUpdate:
      type: object
      properties:
        status:
          type: string
        ambulanceId:
          type: string
        hospitalId:
          type: string
        responderNotes:
          type: string
      required: [status]
    VitalsReading:
      type: object
      properties:
        id:
          type: string
        incidentId:
          type: string
        heartRate:
          type: integer
        temperature:
          type: number
        accelerometerX:
          type: ["number", "null"]
        accelerometerY:
          type: ["number", "null"]
        accelerometerZ:
          type: ["number", "null"]
        impactMagnitude:
          type: ["number", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        confidence:
          type: number
        alerts:
          type: array
          items:
            type: string
        timestamp:
          type: string
      required: [id, incidentId, heartRate, temperature, confidence, alerts, timestamp]
    VitalsInput:
      type: object
      properties:
        heartRate:
          type: integer
        temperature:
          type: number
        accelerometerX:
          type: number
        accelerometerY:
          type: number
        accelerometerZ:
          type: number
        latitude:
          type: number
        longitude:
          type: number
        confidence:
          type: number
      required: [heartRate, temperature, confidence]
...
    Ambulance:
      type: object
      properties:
        id:
          type: string
        unitNumber:
          type: string
        driverName:
          type: string
        driverPhone:
          type: string
        status:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        speedKmh:
          type: ["number", "null"]
        assignedIncidentId:
          type: ["string", "null"]
        updatedAt:
          type: string
      required: [id, unitNumber, driverName, driverPhone, status, latitude, longitude, updatedAt]
    LocationUpdate:
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
        speedKmh:
          type: number
      required: [latitude, longitude]
...
    SensorDataInput:
      type: object
      properties:
        deviceId:
          type: string
        accelerometerX:
          type: number
        accelerometerY:
          type: number
        accelerometerZ:
          type: number
        heartRate:
          type: integer
        skinTemperature:
          type: number
        latitude:
          type: number
        longitude:
          type: number
        timestamp:
          type: string
      required: [deviceId, accelerometerX, accelerometerY, accelerometerZ, timestamp]
    SensorDataResult:
      type: object
      properties:
        impactDetected:
          type: boolean
        severity:
          type: string
        impactMagnitude:
          type: number
        filteredHeartRate:
          type: ["integer", "null"]
        filteredTemperature:
          type: ["number", "null"]
        incidentCreated:
          type: boolean
        incidentId:
          type: ["string", "null"]
        alerts:
          type: array
          items:
            type: string
      required: [impactDetected, severity, impactMagnitude, incidentCreated, alerts]
...
    NearestAmbulanceResult:
      type: object
      properties:
        ambulance:
          $ref: "#/components/schemas/Ambulance"
        distanceKm:
          type: number
        etaMinutes:
          type: number
        bearing:
          type: number
        allAmbulances:
          type: array
          items:
            type: object
            additionalProperties: true
      required: [ambulance, distanceKm, etaMinutes]
...