# lib/api-zod/src/generated/types/nearestAmbulanceResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RoadSoS Kids Emergency Response System API
 * OpenAPI spec version: 0.1.0
 */
import type { Ambulance } from './ambulance';
import type { NearestAmbulanceResultAllAmbulancesItem } from './nearestAmbulanceResultAllAmbulancesItem';

export interface NearestAmbulanceResult {
  ambulance: Ambulance;
  distanceKm: number;
  etaMinutes: number;
  bearing?: number;
  allAmbulances?: NearestAmbulanceResultAllAmbulancesItem[];
}